[LOG-4732] Address recursive discovery review feedback
