+++ src/core/scope.ts
@@ -1,0 +1,6 @@
+/**
+ * Install-scope domain type (§5.2, §7.4).
+ */
+
+/** Install scope: user-wide or project-local. */
+export type Scope = "user" | "project";
