+++ src/install/resolve/dep.ts
@@ -8,12 +8,10 @@ import { acquireTap } from "../../sources/acquire/index.ts";
 import type { SkippedSkill } from "../../sources/expand.ts";
 import { findSiblingDep, type SiblingHit } from "../dep-resolution.ts";
 import { attributeRef } from "../tap-attribution.ts";
+import { expandSkillsAsItems, sourcePinned, sourceRequestedRef } from "./expand-items.ts";
 import {
   enqueueTapRef,
-  expandSkillsAsItems,
   type PendingItem,
-  sourcePinned,
-  sourceRequestedRef,
 } from "./enqueue.ts";
 
 /** Resolve and enqueue items for a dependency reference. */
