+++ src/install/resolve/enqueue.ts
@@ -2,20 +2,15 @@
  * Enqueue helpers for the install resolver (§9).
  *
  * `resolveInstallSet` in `./index.ts` orchestrates the dependency
- * walk; the per-reference work (parsing, expansion, dependency-edge
- * handling) lives here to keep each file under the 200-line cap.
- *
- * Exports:
- *   - `PendingItem` — an in-flight entry the orchestrator holds.
- *   - `enqueueTapRef` — turn a `TapSource` into zero+ PendingItems.
- *   - `expandSkillsAsItems` — map one directory to PendingItems.
- *   - `sourceRequestedRef` / `sourcePinned` — structural helpers.
+ * walk; tap-reference expansion lives here to keep that orchestration
+ * separate from reference resolution details.
  */
 
-import type { Config, LoadedSkill, Source, TapConfig } from "../../core/types.ts";
+import type { Config, LoadedSkill, TapConfig } from "../../core/types.ts";
 import { acquireTap } from "../../sources/acquire/index.ts";
-import { expandSkills, type SkippedSkill } from "../../sources/expand.ts";
+import type { SkippedSkill } from "../../sources/expand.ts";
 import { type KindHint, resolveTapRef } from "../resolve-ref/index.ts";
+import { expandSkillsAsItems } from "./expand-items.ts";
 
 export interface PendingItem {
   readonly loaded: LoadedSkill;
@@ -38,12 +33,9 @@ export function enqueueTapRef(
   kindHint: KindHint,
 ): { items: PendingItem[]; config: Config; skipped: readonly SkippedSkill[] } {
   // Whole-tap install short-circuit: bare `<name>` that is the name of
-  // a configured tap. Cross-tap collisions (a same-named skill in a
-  // DIFFERENT tap) are handled earlier in the CLI layer via
-  // `detectCollision`; reaching this point means the CLI already
-  // prompted or `--yes` was in play. Within-tap collisions (the same
-  // name also being a skill in the matched tap) are rare and we
-  // preserve the legacy "tap wins" behavior for back-compat.
+  // a configured tap. Cross-tap collisions are handled earlier in the
+  // CLI layer via `detectCollision`; reaching this point means the CLI
+  // already prompted or `--yes` was in play.
   if (
     source.tap === null &&
     source.namespace === null &&
@@ -66,8 +58,6 @@ export function enqueueTapRef(
     }
   }
 
-  // Everything else runs through the resolver, which produces a
-  // concrete NameCandidate (skill / namespace / tap) or throws.
   const candidate = resolveTapRef(
     {
       type: "tap",
@@ -102,9 +92,6 @@ export function enqueueTapRef(
     return { items, config, skipped };
   }
 
-  // kind === "skill". The short-circuit above handles kind === "tap"
-  // for bare names before we reach the resolver; other paths
-  // (2-seg, 3-seg) never resolve to a tap.
   const skill = candidate as Extract<typeof candidate, { kind: "skill" }>;
   const acquired = acquireTap(skill.tap, home);
   const expansion = expandSkillsAsItems(
@@ -118,61 +105,4 @@ export function enqueueTapRef(
     false,
   );
   return { items: expansion.items, config, skipped: expansion.skipped };
-}
-
-interface ExpansionItems {
-  readonly items: PendingItem[];
-  readonly skipped: readonly SkippedSkill[];
-}
-
-/** Walk `dir` and produce one PendingItem per skill found (single or expanded). */
-export function expandSkillsAsItems(
-  dir: string,
-  tap: TapConfig,
-  baseTapPath: string,
-  resolvedSha: string | null,
-  requestedRef: string | null,
-  pinned: boolean,
-  explicit: boolean,
-  tracksTap: boolean,
-): ExpansionItems {
-  const { valid, skipped } = expandSkills(dir, { recursive: tap.discovery === "recursive" });
-  const items: PendingItem[] = [];
-  for (const l of valid) {
-    // tap-relative path of this skill's directory inside its tap.
-    const subSegment = l.path === dir ? "" : l.path.slice(dir.length + 1);
-    const tapRelativePath = baseTapPath
-      ? subSegment
-        ? `${baseTapPath}/${subSegment}`
-        : baseTapPath
-      : subSegment;
-    items.push({
-      loaded: l,
-      tap,
-      tapRelativePath,
-      resolvedSha,
-      requestedRef,
-      pinned,
-      explicit,
-      tracksTap,
-    });
-  }
-  return { items, skipped };
-}
-
-export function sourceRequestedRef(source: Source): string | null {
-  if (source.type === "git") return source.ref;
-  if (source.type === "tap") return source.ref;
-  return null;
-}
-
-export function sourcePinned(source: Source, resolvedSha: string | null): boolean {
-  if (source.type !== "git") return false;
-  if (source.ref === null) return false;
-  // Without talking to git here (resolveRef happens in acquireTap and
-  // doesn't currently classify), assume any explicit ref counts as
-  // pinned. resolvedSha is unused in the heuristic but kept in the
-  // signature for future use.
-  void resolvedSha;
-  return true;
 }
